Move useful ast building functions into their own module.
